requirements.txt
# requirements.txt
streamlit
python-dotenv
pandas
matplotlib
seaborn
langchain
langchain-openai
langchain-google-genai
langchain-community
langgraph
langchain-chroma
pypdf
streamlit